Move OdbcRowWriter from api/ into its own row/ module
