+++ mssql-odbc/src/lib.rs
@@ -9,6 +9,7 @@ mod connection;
 mod error;
 mod handles;
 mod params;
+pub(crate) mod row;
 #[cfg(test)]
 pub(crate) mod test_support;
